@@ -7,7 +7,7 @@ buildscript {
         maven { url = 'https://repo.spongepowered.org/maven/' }
     }
     dependencies {
-        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '3.+', changing: true
+        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.+', changing: true
         classpath 'org.spongepowered:mixingradle:0.7-SNAPSHOT'
     }
 }
@@ -15,14 +15,16 @@ apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'eclipse'
 apply plugin: 'maven-publish'
 
-version = '2.2.0'
+version = '3.2.0'
 group = 'com.chocohead.mm' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'mm'
 
-sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8'
+sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '17'
+//sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8'
 
 minecraft {
-    mappings channel: 'snapshot', version: '20210309-1.16.5'
+    mappings channel: 'official', version: '1.20.1'
+    //mappings channel: 'snapshot', version: '20210309-1.16.5'
     accessTransformer = file('src/main/resources/META-INF/accesstransformer.cfg')
 
     runs {
@@ -74,9 +76,10 @@ repositories {
 }
 
 dependencies {
-    minecraft 'net.minecraftforge:forge:1.16.5-36.1.0'
+    minecraft 'net.minecraftforge:forge:1.20.1-47.1.3'
+    //minecraft 'net.minecraftforge:forge:1.16.5-36.1.0'
     implementation 'org.spongepowered:noise:2.0.0-SNAPSHOT'
-    compile 'org.spongepowered:mixin:0.8.1'
+    //compile 'org.spongepowered:mixin:0.8.1'
 }
 
 jar {
